refactor: consolidate duplicated chain utils into @gridplus/chain-core
Move toBuffer, parseHexBytes, buildSigResultFromRsv, and
compressSecp256k1Pubkey into chain-core and re-export from each
chain's devices/shared.ts so consumer imports stay unchanged.
