+++ packages/chains/cosmos/src/chain.ts
@@ -11,6 +11,7 @@ import type {
   SignResult,
   Signer as CoreSigner,
 } from '@gridplus/chain-core';
+import { compressSecp256k1Pubkey } from '@gridplus/chain-core';
 import { bech32 } from 'bech32';
 import { ripemd160 } from '@noble/hashes/ripemd160';
 import { sha256 } from '@noble/hashes/sha256';
@@ -80,23 +81,6 @@ const buildPath = (
     addressIndex,
   ];
 };
-
-function compressSecp256k1Pubkey(pubkey: Uint8Array): Uint8Array {
-  if (pubkey.length === 33 && (pubkey[0] === 0x02 || pubkey[0] === 0x03)) {
-    return pubkey;
-  }
-  if (pubkey.length === 65 && pubkey[0] === 0x04) {
-    const x = pubkey.slice(1, 33);
-    const yLastByte = pubkey[64];
-    const prefix = yLastByte % 2 === 0 ? 0x02 : 0x03;
-    const out = new Uint8Array(33);
-    out[0] = prefix;
-    out.set(x, 1);
-    return out;
-  }
-  // Unknown format, return as-is.
-  return pubkey;
-}
 
 const pubkeyToBech32Address = (pubkeyCompressed: Uint8Array, hrp: string) => {
   const digest = ripemd160(sha256(pubkeyCompressed));
